Add additionalDocumentsCheckAnswers.ts and riskInformationCheckAnswers.ts
